Updated assignment_grades and create_grade
Assignment_grades shows all students and their grades for an assignment if any. If there is a grade for an assignment, URLs for editing and deleting are shown; if not, a URL for creating a grade entry is shown.

Currently create_grade redirects to the base Assignment view after a successful submission. Attempting to just return to create_grade somehow drops a parameter from the URL string.